+++ source/plugins/go/src/utils.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"Docker-Provider/source/plugins/go/src/extension"
 	"bufio"
 	"crypto/tls"
 	"errors"
@@ -279,4 +280,31 @@ func getGenevaWindowsNamedPipeName() string {
 		namedPipeName = "CAgentStream_ContainerLogV2Pipe_" + gcsNameSpace
 	}
 	return namedPipeName
+}
+
+// get the Output stream ID tag value corresponding to the datatype
+func getOutputStreamIdTag(dataType string) string {
+	useFromCache := true
+	switch dataType {
+	case ContainerLogDataType, ContainerLogV2DataType:
+		elapsed := time.Now().Sub(MdsdContainerLogTagRefreshTracker)
+		if !strings.HasPrefix(MdsdContainerLogTagName, MdsdOutputStreamIdTagPrefix) || elapsed.Seconds() >= agentConfigRefreshIntervalSeconds {
+			useFromCache = false
+			MdsdContainerLogTagRefreshTracker = time.Now()
+		}
+	case KubeMonAgentEventDataType:
+		elapsed := time.Now().Sub(MdsdKubeMonAgentEventsTagRefreshTracker)
+		if !strings.HasPrefix(MdsdKubeMonAgentEventsTagName, MdsdOutputStreamIdTagPrefix) || elapsed.Seconds() >= agentConfigRefreshIntervalSeconds {
+			useFromCache = false
+			MdsdKubeMonAgentEventsTagRefreshTracker = time.Now()
+		}
+	case InsightsMetricsDataType:
+		elapsed := time.Now().Sub(MdsdInsightsMetricsTagRefreshTracker)
+		if !strings.HasPrefix(MdsdInsightsMetricsTagName, MdsdOutputStreamIdTagPrefix) || elapsed.Seconds() >= agentConfigRefreshIntervalSeconds {
+			useFromCache = false
+			MdsdInsightsMetricsTagRefreshTracker = time.Now()
+		}
+	}
+
+	return extension.GetInstance(FLBLogger, ContainerType).GetOutputStreamId(dataType, useFromCache)
 }
